requirements.txt
google-genai==1.12.1
python-dotenv==1.1.0
requests
